Rename component file
