+++ src/components/CalendarWeekHeader.vue
@@ -1,6 +1,8 @@
 <template>
   <th>
-    <p>{{ day }}</p>
+    <div class="week-header">
+      {{ day }}
+    </div>
   </th>
 </template>
 
@@ -15,6 +17,8 @@ export default {
 }
 </script>
 
-<style>
-
+<style scoped>
+.week-header {
+  margin-top: 15px;
+}
 </style>
